# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DEV-001 Thyroid Pre-Visit Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: loginStaff
      tags: [auth]
      summary: Staff login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshInput"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRefreshResult"
  /auth/logout:
    post:
      operationId: logoutStaff
      tags: [auth]
      summary: Staff logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current staff user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Patient Auth ──────────────────────────────────────────────────────────
...
  /doctor/appointments/{id}/summaries:
    get:
      operationId: getAppointmentSummaries
      tags: [doctor]
      summary: Get templated summaries for appointment (doctor)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummariesResult"
  /doctor/appointments/{id}/documents:
    get:
      operationId: getDoctorDocuments
      tags: [doctor]
      summary: List documents for appointment (doctor)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentsResult"
  /doctor/appointments/{id}/documents/{docId}/download:
    get:
      operationId: getDocumentDownload
      tags: [doctor]
      summary: Get signed download URL for document (doctor)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: docId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Download info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDownload"
  /doctor/patients/{patientId}/history:
    get:
      operationId: getPatientHistory
      tags: [doctor]
      summary: Patient appointment history (doctor)
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientHistory"

  # ── Admin ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    patientAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ────────────────────────────────────────────────────────────────
    AuthCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        mfaToken:
          type: ["string", "null"]
    StaffSession:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: ["string", "null"]
        requiresMfa:
          type: boolean
        user:
          $ref: "#/components/schemas/StaffUser"
    TokenRefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    TokenRefreshResult:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
    StaffUser:
      type: object
      required: [id, email, role, fullName, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [doctor, clinic_admin, nurse]
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        mfaEnabled:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
...
    Appointment:
      type: object
      required: [id, invitedFullName, invitedPhone, appointmentType, scheduledAt, status, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: ["string", "null"]
        doctorId:
          type: string
        invitedFullName:
          type: string
        invitedPhone:
          type: string
        appointmentType:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft_invitation, link_sent, opened, in_progress, submitted, locked, reopened, rescheduled, cancelled]
        labStatus:
          type: ["string", "null"]
        excludedFromClinicalViews:
          type: boolean
        originalScheduledAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        doctor:
          $ref: "#/components/schemas/StaffUser"
        patient:
          $ref: "#/components/schemas/Patient"
...
    Patient:
      type: object
      required: [id, fullName, phone, dateOfBirth, createdAt]
      properties:
        id:
          type: string
        fullName:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        sex:
          type: ["string", "null"]
        heightCm:
          type: ["string", "null"]
        matchStatus:
          type: string
        duplicateReviewFlag:
          type: boolean
        consentGivenAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    PatientHistory:
      type: object
      required: [patient, appointments]
      properties:
        patient:
          $ref: "#/components/schemas/Patient"
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"

    # ── Questionnaires ──────────────────────────────────────────────────────
...
    Document:
      type: object
      required: [id, appointmentId, originalFileName, mimeType, fileSizeBytes, createdAt]
      properties:
        id:
          type: string
        appointmentId:
          type: string
        originalFileName:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: integer
        documentType:
          type: ["string", "null"]
        labStatus:
          type: ["string", "null"]
        deletedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    DocumentsResult:
      type: object
      required: [documents]
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
    DocumentDownload:
      type: object
      required: [fileName, mimeType]
      properties:
        downloadUrl:
          type: ["string", "null"]
        fileName:
          type: string
        mimeType:
          type: string

    # ── Doctor Clinical ─────────────────────────────────────────────────────
    Summary:
      type: object
      required: [id, appointmentId, variant, content, generatedAt, templateVersion]
      properties:
        id:
          type: string
        appointmentId:
          type: string
        variant:
          type: string
          enum: [current_visit, current_visit_plus_history]
        content:
          type: string
        generatedAt:
          type: string
          format: date-time
        templateVersion:
          type: string
    SummariesResult:
      type: object
      required: [summaries]
      properties:
        summaries:
          type: array
          items:
            $ref: "#/components/schemas/Summary"

    # ── Admin ───────────────────────────────────────────────────────────────
...