# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DEV-001 Thyroid Pre-Visit Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthCredentials {
  email: string;
  password: string;
  /** @nullable */
  mfaToken?: string | null;
}